Refactor the code to enhance readability
