+++ streamlit_langgraph/utils/__init__.py
@@ -3,6 +3,7 @@
 from .file_handler import FileHandler, MIME_TYPES
 from .custom_tool import CustomTool
 from .mcp_tool import MCPToolManager
+from .message_utils import create_message_with_id
 
 __all__ = [
     # File handling
@@ -12,4 +13,6 @@ __all__ = [
     "CustomTool",
     # MCP tools
     "MCPToolManager",
+    # Message utilities
+    "create_message_with_id",
 ]
