+++ streamlit_langgraph/workflow/agent_nodes/__init__.py
@@ -1,7 +1,7 @@
 # Agent node creation modules for streamlit-langgraph.
 
-from .factory import AgentNodeFactory
-from .base import AgentNodeBase, create_message_with_id
+from .factory import AgentNodeFactory, AgentNodeBase
+from ...utils import create_message_with_id
 from .handoff_delegation import HandoffDelegation
 from .tool_calling_delegation import ToolCallingDelegation
