+++ streamlit_langgraph/workflow/patterns/hierarchical.py
@@ -9,6 +9,7 @@ from langgraph.graph import StateGraph, START, END
 
 from ...agent import Agent
 from ..agent_nodes.factory import AgentNodeFactory
+from ..agent_nodes.routing import RoutingHelper
 from ...core.state import WorkflowState
 
 
@@ -67,20 +68,17 @@ class HierarchicalPattern:
         
         graph = StateGraph(WorkflowState)
         
-        # Extract all sub-supervisors for top-level delegation
+        # Setting up top supervisor node
         sub_supervisors = [team.supervisor for team in supervisor_teams]
-        
-        # Top supervisor treats sub-supervisors as its "workers"
-        # Hierarchical pattern doesn't support parallel at this level
         top_supervisor_node = AgentNodeFactory.create_supervisor_agent_node(
             top_supervisor, sub_supervisors, allow_parallel=False
         )
         graph.add_node(top_supervisor.name, top_supervisor_node)
         graph.add_edge(START, top_supervisor.name)
         
+        # Setting up sub-supervisor nodes and worker nodes
+        # sub-supervisor is just a supervisor for their team\
         for team in supervisor_teams:
-            # Sub-supervisor is just a supervisor for their team
-            # Hierarchical pattern doesn't support parallel at team level
             sub_supervisor_node = AgentNodeFactory.create_supervisor_agent_node(
                 team.supervisor, team.workers, allow_parallel=False
             )
@@ -93,9 +91,7 @@ class HierarchicalPattern:
                 graph.add_node(worker.name, worker_node)
         
         # Add routing using supervisor pattern routing logic
-        graph = HierarchicalPattern._add_hierarchical_routing(
-            graph, top_supervisor, supervisor_teams
-        )
+        graph = HierarchicalPattern._add_hierarchical_routing(graph, top_supervisor, supervisor_teams)
         
         return graph.compile(checkpointer=workflow_checkpointer)
     
@@ -113,18 +109,7 @@ class HierarchicalPattern:
         """
         
         sub_supervisor_names = [team.supervisor.name for team in supervisor_teams]
-        
-        # Top supervisor routing (same pattern as supervisor -> workers)
-        def top_supervisor_route(state: WorkflowState) -> str:
-            """Top supervisor routes to sub-supervisors or END."""
-            routing_decision = state["metadata"].get("routing_decision", {})
-            action = routing_decision.get("action", "finish")
-            
-            if action == "delegate":
-                target = routing_decision.get("target_worker", "")
-                if target in sub_supervisor_names:
-                    return target
-            return "__end__"
+        top_supervisor_route = RoutingHelper.create_hierarchical_top_route(sub_supervisor_names)
         
         top_routes = {name: name for name in sub_supervisor_names}
         top_routes["__end__"] = END
@@ -133,28 +118,16 @@ class HierarchicalPattern:
         # For each team, apply standard supervisor pattern routing
         for team in supervisor_teams:
             worker_names = [w.name for w in team.workers]
-            # Sub-supervisor routes to workers or back to top supervisor
-            # Same logic as supervisor routing, but returns to top supervisor instead of END
-            def make_subsupervisor_route(team_worker_names, top_sup_name):
-                def subsupervisor_route(state: WorkflowState) -> str:
-                    """Sub-supervisor routes to workers or back to top supervisor."""
-                    routing_decision = state["metadata"].get("routing_decision", {})
-                    action = routing_decision.get("action", "finish")
-                    
-                    if action == "delegate":
-                        target = routing_decision.get("target_worker", "")
-                        if target in team_worker_names:
-                            return target
-                    # When done, return to top supervisor (not END)
-                    return top_sup_name
-                return subsupervisor_route
+            subsupervisor_route = RoutingHelper.create_hierarchical_sub_route(
+                worker_names, top_supervisor.name
+            )
             
             sub_routes = {name: name for name in worker_names}
             sub_routes[top_supervisor.name] = top_supervisor.name
             
             graph.add_conditional_edges(
                 team.supervisor.name,
-                make_subsupervisor_route(worker_names, top_supervisor.name),
+                subsupervisor_route,
                 sub_routes
             )
